requirements.txt
numpy~=1.21.6
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
tensorflow-metal==0.5.0
keras==2.9.0
pandas~=1.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.6
Pillow==9.2.0
